move calculate_n_bins function to conversion
