+++ elephant/test/test_conversion.py
@@ -762,6 +762,36 @@ class DiscretiseSpiketrainsTestCase(unittest.TestCase):
         np.testing.assert_array_equal(bsts.to_array(),
                                       bsts_discretised.to_array())
 
+class CalculateNBinsTestCase(unittest.TestCase):
+    def test_calculate_n_bins_not_time_quantity(self):
+        # t_start is in Hz
+        self.assertRaises(TypeError, cv.calculate_n_bins,
+                          t_start=1*pq.Hz, t_stop=5*pq.ms, bin_size=1*pq.ms)
+
+    def test_calculate_n_bins_different_units(self):
+        self.assertWarns(UserWarning, cv.calculate_n_bins,
+                         t_start=1*pq.s, t_stop=5*pq.ms, bin_size=1*pq.ms)
+
+    def test_calculate_n_bins_int(self):
+        n_bins = cv.calculate_n_bins(t_start=0*pq.s, t_stop=10*pq.ms,
+                                        bin_size=3*pq.ms)
+        self.assertIsInstance(n_bins, int)
+        self.assertEqual(n_bins, 3)
+
+    def test_calculate_n_bins_float(self):
+        n_bins = cv.calculate_n_bins(t_start=0.0*pq.s,
+                                             t_stop=9.0*pq.ms,
+                                             bin_size=3.0*pq.ms)
+        self.assertIsInstance(n_bins, int)
+        self.assertEqual(n_bins, 3)
+
+    def test_calculate_n_bins_rounding(self):
+        n_bins = cv.calculate_n_bins(t_start=0*pq.s,
+                                        t_stop=0.99999999*pq.ms,
+                                        bin_size=1.0*pq.ms)
+        self.assertIsInstance(n_bins, int)
+        self.assertEqual(n_bins, 1)
+
 
 if __name__ == '__main__':
     unittest.main()
